Adding automatic documentation generator
