+++ docs/Doxyfile
@@ -783,7 +783,8 @@ WARN_LOGFILE           =
 INPUT                  = ../include/goofit \
                          ../examples \
                          ../src \
-                         ../docs/documentation.md
+                         ../docs/documentation.md \
+                         ../docs/pdfs
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
@@ -812,7 +813,8 @@ FILE_PATTERNS          = *.cu \
                          *.h \
                          *.cpp \
                          *.hpp \
-                         *.cc
+                         *.cc \
+                         *.md
 
 # The RECURSIVE tag can be used to specify whether or not subdirectories should
 # be searched for input files as well.
