+++ python/PDFs/basic/ArgusPdf.cpp
@@ -1,27 +1,12 @@
-#include <pybind11/pybind11.h>
+#include <goofit/Python.h>
 
 #include <goofit/PDFs/basic/ArgusPdf.h>
 #include <goofit/Variable.h>
+#include <goofit/docs/ArgusPdf.h>
 
 using namespace GooFit;
-namespace py = pybind11;
-using namespace pybind11::literals;
-
-struct HelpPrinter {
-    std::string help_str;
-    std::string getHelp() const { return help_str; }
-    HelpPrinter(std::string input)
-        : help_str(input) {}
-};
 
 void init_ArgusPdf(py::module &m) {
-    py::class_<HelpPrinter>(m, "HelpPrinter")
-        .def(py::init<std::string>())
-        .def("_repr_markdown_", &HelpPrinter::getHelp)
-        .def("__repr__", &HelpPrinter::getHelp);
-
-    m.attr("ArgusHelp") = HelpPrinter(ArgusHelp);
-
     py::class_<ArgusPdf, GooPdf>(m, "ArgusPdf")
         .def(py::init<std::string, Observable, Variable, Variable, bool>(),
              "Standard Argus",
@@ -39,5 +24,6 @@ void init_ArgusPdf(py::module &m) {
              "s"_a,
              "upper"_a,
              "power"_a)
-        .def_static("help", []() { return HelpPrinter(ArgusHelp); });
+
+        .def_static("help", []() { return HelpPrinter(ArgusPdf_docs); });
 }
