+++ python/include/goofit/PyProps.h
